% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis.R
\name{around}
\alias{around}
\title{Print a data frame row and its neighbors}
\usage{
around(df, row, d = 2)
}
\arguments{
\item{df}{the data frame}

\item{row}{the row index}
}
\description{
Print a data frame row and its neighbors
}

