\name{cr-package}
\alias{cr-package}
\alias{cr}
\docType{package}
\title{
Miscellaneous R scripts packaged for Clark Richards.
}
\description{
Miscellaneous R scripts packaged for Clark Richards. Contains some code
not written by me, so it is not meant for wider distribution.
}
\details{
\tabular{ll}{
Package: \tab cr\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2011-09-27\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
Clark Richards

Maintainer: <clark.richards@gmail.com>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[oce:oce-package]{oce}} ~~
}
\examples{
}
