% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_format_pval.R
\name{cr_format.pval}
\alias{cr_format.pval}
\title{Format P-values for publication}
\usage{
cr_format.pval(x)
}
\arguments{
\item{x}{a numeric value (= a p-value in numeric form)}
}
\value{
A character string.
}
\description{
Formats your numeric p-values to a truncated character string fit for publication.
For instance; 0.00000431 will be returned as "<0.001", and 0.005 as "0.006".
}
