% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stringtools.R
\name{lowercase}
\alias{lowercase}
\title{Converts a string to lowercase based on index}
\usage{
lowercase(str, index = NULL)
}
\description{
Converts a string to lowercase based on index
}
\examples{
lowercase("Dr. Dre", 1)
lowercase("Dr. Dre")
}

