% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhia_ordlista.R
\docType{data}
\name{rhia_ordlista}
\alias{rhia_ordlista}
\title{rhia_ordlista}
\format{
A named list with 299 items.
}
\source{
\url{https://www.ucr.uu.se/swedeheart/dokument-sh/variabellista}
}
\usage{
rhia_ordlista
}
\description{
Conversion list for RIKS-HIA variable levels

A list containing a named vector for each RIKS-HIA variable. 
The vector can be accessed subsetting the list on variable name. 
Note that all variable names is in lowercase
}
\keyword{datasets}
