% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stan.R
\name{run_model}
\alias{run_model}
\title{Run the mode}
\usage{
run_model(path, party = NULL, party_save_str = NULL, chains = 1, iter,
  seed = sample(1:100, length(chains)))
}
\description{
Run the mode
}

