% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhia_ordlista.R
\docType{data}
\name{sephia_ordlista}
\alias{sephia_ordlista}
\title{sephia_ordlista}
\format{
A named list
}
\source{
\url{https://www.ucr.uu.se/swedeheart/dokument-sh/variabellista}
}
\usage{
sephia_ordlista
}
\description{
Conversion list for SEPHIA variable levels

A list containing a named vector for each SEPHIA variable. 
The vector can be accessed subsetting the list on variable name. 
Note that all variable names is in lowercase
}
\keyword{datasets}
