% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhia_ordlista.R
\docType{data}
\name{sh_gem_ordlista}
\alias{sh_gem_ordlista}
\title{sh_gem_ordlista}
\format{
A named list
}
\source{
\url{https://www.ucr.uu.se/swedeheart/dokument-sh/variabellista}
}
\usage{
sh_gem_ordlista
}
\description{
Conversion list for SWEDEHEART GEMENSAM variable levels

A list containing a named vector for each SWEDEHEART GEMENSAM variable. 
The vector can be accessed subsetting the list on variable name. 
Note that all variable names is in lowercase
}
\keyword{datasets}
