% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stan.R
\name{stan2coda}
\alias{stan2coda}
\title{Converts stan to coda object}
\usage{
stan2coda(fit)
}
\description{
Converts stan to coda object
}
\details{
Adopted from Jeromy Anglim (http://jeromyanglim.tumblr.com/post/91434443911/how-to-convert-a-stan-fit-object-to-work-with-coda)
}

