% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survexp.R
\name{table_survexp}
\alias{table_survexp}
\title{Convert survex to survfit object}
\usage{
table_survexp(x)
}
\arguments{
\item{x}{a survexp object from the survexp function}
}
\description{
This function creates a simple data frame containing a survival curve from a survex object.
}
