% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stringtools.R
\name{take}
\alias{take}
\title{Takes elements from a character vector based on Regular Expression}
\usage{
take(str, re)
}
\description{
Takes elements from a character vector based on Regular Expression
}
\examples{
vec <- c("01", "02", "3", "4", "05")
take(vec, "^0")
}

