% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot (cc's conflicted copy 2015-08-12).R, R/ggplot.R
\name{theme_cr}
\alias{theme_cr}
\title{Theme by Christian Rubba}
\usage{
theme_cr(base_size = 12, base_family = "sans")

theme_cr(base_size = 12, base_family = "sans")
}
\description{
Theme by Christian Rubba

Theme by Christian Rubba
}

