% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wernickethemes.R
\name{theme_wernicke_base}
\alias{theme_wernicke_base}
\title{theme_wernicke_base()}
\usage{
theme_wernicke_base()
}
\description{
the base for my personal ggplot themes. This one is not very interesting on its own, but can be used as a foundation for theme customing.
}
\examples{
p + theme_wernicke_grey()

}
