% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wernickethemes.R
\name{theme_wernicke_dark}
\alias{theme_wernicke_dark}
\title{theme_wernicke_dark()}
\usage{
theme_wernicke_dark()
}
\description{
A dark version of my theme. Does not work very well yet, but I am sure it has potential.
}
\examples{
p + scale_fill_viridis_d() + theme_wernicke_dark()
p + scale_fill_manual(values = wes_palette("FantasticFox1")) + theme_wernicke_dark()

}
