% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wernickethemes.R
\name{theme_wernicke_grey}
\alias{theme_wernicke_grey}
\title{theme_wernicke_grey()}
\usage{
theme_wernicke_grey()
}
\description{
A light grey version of my theme. Works well with viridis() colors
}
\examples{
p + scale_fill_viridis_d() + theme_wernicke_grey()

}
