% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wernickethemes.R
\name{theme_wernicke_white}
\alias{theme_wernicke_white}
\title{theme_wernicke_white()}
\usage{
theme_wernicke_white()
}
\description{
A version of my theme with white background. Looks nice with viridis and magma colors.
}
\examples{
p + scale_fill_viridis_d() + theme_wernicke_white()
p + scale_fill_manual(values = wes_palette("Zissou1")) + theme_wernicke_white()

}
