% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stringtools.R
\name{uppercase}
\alias{uppercase}
\title{Converts a string to uppercase based on index}
\usage{
uppercase(str, index = NULL)
}
\description{
Converts a string to uppercase based on index
}
\examples{
uppercase("Dr. Dre", 2)
uppercase("Dr. Dre", -1)
}

