% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wernickethemes.R
\name{wer_classic}
\alias{wer_classic}
\title{wer_classic()}
\usage{
wer_classic(
  base_size = 11,
  base_family = "Gill Sans",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\description{
A classic wernicke theme
}
