% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wernickethemes.R
\name{wesanderson_color}
\alias{wesanderson_color}
\title{wesanderson_color()}
\usage{
wesanderson_color(string_color, ...)
}
\arguments{
\item{string_color}{A string corresponding to a color palette from the wesanderson package.}
}
\description{
Convenience function for using a wesanderson package color palette.
}
\examples{
p + theme_wernicke_grey() + wesanderson_color("Zissou1")

}
